+++ lib/index.d.ts
@@ -7,6 +7,16 @@ export interface StoreObjectConfig<O extends Record<string, any>> {
      */
     checkGets?: boolean;
     /**
+     * Whether the stored object only contains/stores *some* of the keys on the serialized object.
+     * This is useful if you want an object to look at only some keys of a localStorage object
+     * without overwriting the other ones.
+     *
+     * It's important to note that passing this option effectively enables a key validation of sorts;
+     * any keys that were not passed are ignored and not passed to validate or modify (if these methods are defined)
+     * @default false
+     */
+    partial?: boolean;
+    /**
      * Validate an object before setting it in localStorage or reading it.
      * Can confirm/deny if the object is valid, along with an optional error message if it is not
      *
@@ -87,7 +97,7 @@ export interface StoreObjectConfig<O extends Record<string, any>> {
  *
  * @example
  * ```typescript
- * // Validation to automatically change a key based on another
+ * // Automatically change a key based on another
  * import { storeObject } from 'ls-proxy'
  *
  * interface Person {