+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Option to store multiple objects under the same localStorage key (see `partial` parameter for `StoreObjectConfig`
+  or `different-keys-same-object.ts` in examples)
 - Default functions for common validation use cases (key validation and basic type checking).
   See `Validations` namespace